prc.numcu module
